+++ extensions/kommand/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Kommand Changelog
 
+## [Search shortcuts by app and category] - {PR_MERGE_DATE}
+
+- Search shortcuts by app name
+- Search shortcuts by category name
+
 ## [Initial Version] - 2026-03-12
 
 - Add "Show Frontmost App Shortcuts" command